Separate core NAM code from other plugin dsp code and remove IPlug dependency for NAM code.
