+++ NAM/lstm.h
