+++ NAM/numpy_util.cpp
