+++ NAM/util.cpp
