+++ NAM/util.h
